requirements.txt
flask
beautifulsoup4
fuzzysearch
pyOpenSSL